Add embed window test utilities and update related tests
