+++ src/test/helpers.ts
@@ -12,16 +12,6 @@ export type MockConsole = Console & {
     error: ReturnType<typeof vi.fn>;
 };
 
-export type EmbedTestWindowMock = {
-    windowObj: Window;
-    parent: {
-        postMessage: ReturnType<typeof vi.fn>;
-    };
-    listeners: Map<string, (event: MessageEvent) => void>;
-};
-
-const DEFAULT_EMBED_TEST_SEARCH = '?parentOrigin=https%3A%2F%2Fparent.example.com';
-
 export function createMockConsole(): MockConsole {
     return {
         log: vi.fn(),
@@ -30,32 +20,6 @@ export function createMockConsole(): MockConsole {
         warn: vi.fn(),
         error: vi.fn(),
     } as unknown as MockConsole;
-}
-
-export function createEmbedTestWindow(
-    search = DEFAULT_EMBED_TEST_SEARCH,
-    extraWindowProperties: Record<string, unknown> = {},
-): EmbedTestWindowMock {
-    const listeners = new Map<string, (event: MessageEvent) => void>();
-    const parent = {
-        postMessage: vi.fn(),
-    };
-
-    const windowObj = {
-        self: {} as Window,
-        top: {} as Window,
-        parent,
-        location: { search },
-        addEventListener: vi.fn((type: string, handler: (event: MessageEvent) => void) => {
-            listeners.set(type, handler);
-        }),
-        removeEventListener: vi.fn((type: string) => {
-            listeners.delete(type);
-        }),
-        ...extraWindowProperties,
-    } as unknown as Window;
-
-    return { windowObj, parent, listeners };
 }
 
 // Mock Storage implementation
